% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkLearn}
\alias{checkLearn}
\title{Visual checks for the output of \code{tuneLearnFast}}
\usage{
checkLearn(cal)
}
\arguments{
\item{cal}{The output of a call to \code{tuneLearnFast}. This is also contained in the \code{calibr} slot of the output of
a call to \code{mqgam}.}
}
\value{
It provides several plots. The top plot in first page shows the bracket used to estimate log(sigma) for each quantile.
        The brackets are delimited by the crosses and the red dots are the estimates. If a dot falls very close to one of the crosses, 
        that might indicate problems. The bottom plot shows, for each quantile, the value of parameter \code{err} used. Sometimes the algorithm
        needs to increase \code{err} above its user-defined value to achieve convergence. Subsequent plots show, for each quantile, the value
        of the loss function corresponding to each value of log(sigma) explored by Brent algorithm.
}
\description{
Provides so visual checks to verify whether the Brent optimizer used by \code{tuneLearnFast} worked correctly.
}
\examples{
library(qgam); library(MASS)

quSeq <- c(0.4, 0.5, 0.6)
set.seed(737)
fit <- mqgam(accel~s(times, k=20, bs="ad"), data = mcycle, err = 0.01, qu = quSeq, 
             control = list("tol" = 0.01)) # <- semi-sloppy tolerance to speed-up calibration 

checkLearn(fit$calibr)
}
\author{
Matteo Fasiolo <matteo.fasiolo@gmail.com>.
}

