\name{qgam-package}
\alias{qgam-package}
\alias{qgam}
\docType{package}
\title{
\packageTitle{qgam}
}
\description{
\packageDescription{qgam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qgam}
\packageIndices{qgam}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qgam}

Maintainer: \packageMaintainer{qgam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
